Moved map-settings
