+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/map-settings/example-coordinates.tsx
