+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/map-settings/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './map-settings'
+export { default } from './container'
